requirements.txt
click>=6.7
PyYAML==5.1
tqdm>=4.23.4
python-box==5.1.1
flask==1.1.1

# Helpful as dev requirements
jupyter_contrib_nbextensions>=0.5.1
pre-commit==1.14.4
pytest==4.3.1
pytest-cov==2.6.1
sphinx==1.6.6
tox==3.7.0
behave==1.2.6
sphinx-rtd-theme==0.4.3

pandas==1.2.4
schedule==1.1.0
cryptocompare==0.7.5
matplotlib==3.4.1
pykalman==0.9.5
scipy==1.6.3
mplfinance==0.12.7a17
mpl-finance==0.10.1
bayesian-optimization==1.2.0
python-dotenv==0.17.1

